+++ src/main/java/kitchenpos/application/menu/MenuService.java
@@ -2,18 +2,17 @@ package kitchenpos.application.menu;
 
 import kitchenpos.application.menu.request.MenuCreateRequest;
 import kitchenpos.application.menu.request.MenuProductRequest;
-import kitchenpos.domain.menu.MenuRepository;
-import kitchenpos.domain.menu.MenuGroupRepository;
-import kitchenpos.domain.menu.MenuProductRepository;
-import kitchenpos.domain.product.ProductRepository;
 import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.menu.MenuGroupRepository;
 import kitchenpos.domain.menu.MenuProduct;
+import kitchenpos.domain.menu.MenuProductRepository;
+import kitchenpos.domain.menu.MenuRepository;
 import kitchenpos.domain.product.Product;
+import kitchenpos.domain.product.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.math.BigDecimal;
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
