+++ src/main/java/kitchenpos/application/order/OrderService.java
@@ -9,8 +9,8 @@ import kitchenpos.domain.order.OrderDao;
 import kitchenpos.domain.order.OrderLineItem;
 import kitchenpos.domain.order.OrderLineItemRepository;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.order.OrderTable;
-import kitchenpos.domain.order.OrderTableRepository;
+import kitchenpos.domain.tablegroup.OrderTable;
+import kitchenpos.domain.tablegroup.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
