+++ src/main/java/kitchenpos/application/tablegroup/TableGroupService.java
@@ -3,8 +3,8 @@ package kitchenpos.application.tablegroup;
 import kitchenpos.application.tablegroup.request.TableGroupCreateRequest;
 import kitchenpos.domain.order.OrderDao;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.order.OrderTable;
-import kitchenpos.domain.order.OrderTableRepository;
+import kitchenpos.domain.tablegroup.OrderTable;
+import kitchenpos.domain.tablegroup.OrderTableRepository;
 import kitchenpos.domain.tablegroup.TableGroup;
 import kitchenpos.domain.tablegroup.TableGroupRepository;
 import org.springframework.stereotype.Service;
