+++ src/main/java/kitchenpos/domain/tablegroup/OrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain.order;
+package kitchenpos.domain.tablegroup;
 
 import java.util.List;
 import java.util.Optional;
