+++ src/main/java/kitchenpos/domain/tablegroup/TableGroup.java
@@ -1,7 +1,5 @@
 package kitchenpos.domain.tablegroup;
 
-import kitchenpos.domain.order.OrderTable;
-
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
