+++ src/main/java/kitchenpos/ui/order/response/TableResponse.java
@@ -1,6 +1,6 @@
 package kitchenpos.ui.order.response;
 
-import kitchenpos.domain.order.OrderTable;
+import kitchenpos.domain.tablegroup.OrderTable;
 
 public class TableResponse {
