+++ src/test/java/kitchenpos/application/ServiceTestHelper.java
@@ -6,7 +6,7 @@ import kitchenpos.application.menu.request.MenuCreateRequest;
 import kitchenpos.application.menu.request.MenuGroupCreateRequest;
 import kitchenpos.application.menu.request.MenuProductRequest;
 import kitchenpos.application.order.OrderService;
-import kitchenpos.application.order.TableService;
+import kitchenpos.application.tablegroup.TableService;
 import kitchenpos.application.order.request.OrderCreateRequest;
 import kitchenpos.application.order.request.OrderLineItemCreateRequest;
 import kitchenpos.application.order.request.OrderUpdateRequest;
@@ -22,7 +22,7 @@ import kitchenpos.domain.menu.MenuGroup;
 import kitchenpos.domain.menu.MenuProductRepository;
 import kitchenpos.domain.order.Order;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.order.OrderTable;
+import kitchenpos.domain.tablegroup.OrderTable;
 import kitchenpos.domain.product.Product;
 import kitchenpos.domain.tablegroup.TableGroup;
 import org.junit.jupiter.api.BeforeEach;
