+++ src/test/java/kitchenpos/ui/order/OrderAcceptanceTestUtil.java
@@ -9,7 +9,7 @@ import kitchenpos.application.menu.MenuService;
 import kitchenpos.application.menu.request.MenuCreateRequest;
 import kitchenpos.application.menu.request.MenuGroupCreateRequest;
 import kitchenpos.application.menu.request.MenuProductRequest;
-import kitchenpos.application.order.TableService;
+import kitchenpos.application.tablegroup.TableService;
 import kitchenpos.application.order.request.OrderCreateRequest;
 import kitchenpos.application.order.request.OrderLineItemCreateRequest;
 import kitchenpos.application.order.request.OrderUpdateRequest;
@@ -19,7 +19,7 @@ import kitchenpos.application.product.request.ProductCreateRequest;
 import kitchenpos.domain.menu.Menu;
 import kitchenpos.domain.menu.MenuGroup;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.domain.order.OrderTable;
+import kitchenpos.domain.tablegroup.OrderTable;
 import kitchenpos.domain.product.Product;
 import kitchenpos.ui.ControllerAcceptanceTestHelper;
 import kitchenpos.ui.order.response.OrderResponse;
